feat: 도커 환경 가시화 데이터 요청 기능 구현 - api 요청을 보내는 코드 분리   - requestQuizData(), requestVisualizationData() - type을 관리해주는 디렉토리(types) 생성   - Container와 Image 임시 타입 정의
